Refactor main loop into App and service classes
